src/requirements.txt
certifi==2019.11.28
chardet==3.0.4
dataclasses-json==0.3.7
future==0.18.2
idna==2.8
iso8601==0.1.12
marshmallow==3.3.0
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
paho-mqtt==1.5.0
pyserial==3.4
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
stringcase==1.2.0
typing-extensions==3.7.4.1
typing-inspect==0.5.0
urllib3==1.25.7
